Combine RunParallel with Scheduler, create RunLocal
