+++ python/TestHarness/schedulers/Scheduler.py
@@ -7,13 +7,11 @@
 # Licensed under LGPL 2.1, please see LICENSE for details
 # https://www.gnu.org/licenses/lgpl-2.1.html
 import os
-import platform
 import sys
 import threading
 import traceback
 from dataclasses import dataclass
 from enum import Enum
-from importlib.util import find_spec
 from multiprocessing.pool import ThreadPool
 from time import sleep
 from timeit import default_timer as clock
@@ -25,12 +23,11 @@ from FactorySystem.MooseObject import MooseObject
 from TestHarness.JobDAG import JobDAG
 from TestHarness.mpi_config import (
     MPIConfig,
-    MPIType,
     build_hwloc_topology,
     get_mpi_config,
 )
 from TestHarness.StatusSystem import StatusSystem
-from TestHarness.util import findBSDTime, findGNUTime
+from TestHarness.util import findBSDTime, findGNUTime, outputHeader
 
 if TYPE_CHECKING:
     from TestHarness.schedulers.Job import Job
@@ -103,17 +100,15 @@ class Scheduler(MooseObject):
     # This is what will be checked for when we look for valid schedulers
     IS_SCHEDULER = True
 
-    CAN_SET_HWLOC_TOPOLOGY = True
+    CAN_SET_HWLOC_TOPOLOGY = False
     """Whether or not to set hwloc topology if available."""
-    CAN_SET_MAX_MEMORY = True
+    CAN_SET_MAX_MEMORY = False
     """Whether or not Job max memory can be set (Jobs can be killed if over memory)."""
-    CAN_OPENMPI_OVERSUBSCRIBE = True
+    CAN_OPENMPI_OVERSUBSCRIBE = False
     """Whether or not OpenMPI can be set to oversubscribe."""
-    MONITOR_JOB_CPU = True
+    MONITOR_JOB_CPU = False
     """Whether or not this Scheduler should monitor Job process CPU usage."""
-    MONITOR_JOB_MEMORY = (
-        platform.system() != "Windows" and find_spec("psutil") is not None
-    )
+    MONITOR_JOB_MEMORY = False
     """Whether or not this Scheduler should monitor Job process memory usage."""
 
     def __init__(self, harness, params):
@@ -324,8 +319,67 @@ class Scheduler(MooseObject):
         )
 
     def run(self, job):
-        """Call derived run method"""
-        return
+        """Run a tester command"""
+
+        # Build and set the runner that will actually run the commands
+        # This is abstracted away so we can support local runners and PBS/slurm runners
+        job.setRunner(self.buildRunner(job, self.options))
+
+        tester = job.getTester()
+
+        # Do not execute app, and do not processResults
+        if self.options.dry_run:
+            self.setSuccessfulMessage(tester)
+            return
+        # Load results from a previous run
+        elif self.options.show_last_run:
+            job.loadPreviousResults()
+            return
+
+        # Start job timer
+        job.timer.startMain()
+
+        # Anything that throws while running or processing a job should be caught
+        # and the job should fail
+        try:
+            # Launch and wait for the command to finish
+            job.run()
+
+            # Set the successful message
+            if not tester.isSkip() and not job.isFail():
+                self.setSuccessfulMessage(tester)
+        except:
+            trace = traceback.format_exc()
+            job.appendOutput(
+                outputHeader("Python exception encountered in Job") + trace
+            )
+            job.setStatus(job.error, "JOB EXCEPTION")
+        finally:
+            # Stop job timer
+            job.timer.stopMain()
+
+    def setSuccessfulMessage(self, tester):
+        """properly set a finished successful message for tester"""
+        message = ""
+
+        # Handle 'dry run' first, because if true, job.run() never took place
+        if self.options.dry_run:
+            message = "DRY RUN"
+
+        elif tester.specs["check_input"]:
+            message = "SYNTAX PASS"
+
+        elif self.options.scaling and tester.specs["scale_refine"]:
+            message = "SCALED"
+
+        elif (
+            self.options.enable_recover
+            and tester.specs.isValid("skip_checks")
+            and tester.specs["skip_checks"]
+        ):
+            message = "PART1"
+
+        tester.setStatus(tester.success, message)
 
     def augmentJobs(self, jobs):
         """
